requirements.txt
# --- Web Scraping & Trình duyệt ---
selenium==4.18.1
requests>=2.32.2
undetected-chromedriver>=3.5.5

# --- Xử lý Dữ liệu ---
pandas==2.2.1

# --- Trí tuệ Nhân tạo (NLP + ML) ---
underthesea
scikit-learn
transformers==4.38.2
torch==2.2.1
joblib

# --- Tải Video ---
yt-dlp

# --- Web Dashboard ---
streamlit==1.32.0
plotly==5.15.0

python-dotenv==1.0.0